# controlgastos/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Expense tracker API
  description: API for managing budgets and expenses.
  version: v1.0.0
  contact:
    name: Alejandro Granados
    email: alejandro.magb@gmail.com
    url: https://github.com/aledelorbe
...
paths:
...
  /api/budgets:
    get:
      tags:
        - Budgets
      summary: Get the budget
      description: Returns the budget.
      responses:
        '200':
          description: Show the budget
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Budget'

    post:
      tags:
        - Budgets
      summary: Create a budget
      description: Creates a new budget if the sent data is valid.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BudgetInput'
      responses:
        '201':
          description: Budget successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Budget'
        '400':
          description: Validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetValidationErrors'

    delete:
      tags:
        - Budgets
      summary: Delete the budget
      description: Removes the budget record
      responses:
        '200':
          description: The budget deleted successfully
  /api/budgets/{id}:
      put:
        tags:
          - Budgets
        summary: Update an existing budget
        description: Updates all fields of a specific budget by ID.
        parameters:
          - name: id
            in: path
            required: true
            schema:
              type: string
            description: Budget ID
        requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetInput'
        responses:
          '201':
            description: Budget updated successfully.
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Budget'
          '400':
            description: Validation error.
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/BudgetValidationErrors'
          '404':
            description: Budget not found.

  # Expense controller -----------------------------
...
components:
  schemas:
    Budget:
      type: object
      properties:
        id:
          type: string
          example: "65ab12df703e4cbb23a45ed9"
        amount:
          type: number
          format: double
          example: 4500.00
    BudgetInput:
      type: object
      properties:
        amount:
          type: number
          format: double
          example: 4500.00
    BudgetValidationErrors:
      allOf:
        - $ref: "#/components/schemas/ValidationErrorsBase"
      example:
        amount: "The field amount must be equal to or greater than 500"
...
    ValidationErrorsBase:
      type: object
      description: Base validation error schema.
      additionalProperties:
        type: string
...